# WIP: Optimize PK lookup for queries that match exact PK range