--- a/src/Storages/MergeTree/KeyCondition.h
+++ b/src/Storages/MergeTree/KeyCondition.h
@@ -309,6 +309,8 @@ public:
         MonotonicFunctionsChain & functions,
         DataTypePtr current_type);
 
+    bool matchesExactContinuousRange() const;
+
 private:
     /// The expression is stored as Reverse Polish Notation.
     struct RPNElement