@bobrik 2020-07-09
My first version was to allow a prefix of a PK to engage this. E.g. if PK is `(a, b, c)`, then the following is allowed:

* `(a, b, c)`
* `(a, b)`
* `(a)`

But the following isn't:

* `(a, c)`
* `(b, c)`
* `(b)`
* `(c)`

Is that what you mean? If so, I can totally make it happen.