@CurtizJ 2020-07-09
Yes, I meant that.