@bobrik 2020-07-11
I've added support for PK prefix matches and reorganized code a bit to increase code reuse.

Tests should be happy now, the problem was twofold:

* Missing support for monotonic functions — I don't fully understand how to make it work, made it fall back for now.
* Missing pass-by-reference to `may_be_true_in_range()` and final marks, this is fixed.

I've also added some trace logging with amount of matching steps performed, which I intend to add some tests on. Perhaps it can be integrated with `EXPLAIN`, but I think that should be out of scope of this PR.

Let me know what you think.