@canhld94 2022-01-12
Hi @bobrik , sorry for reopening this pr after a long time. This is a nice optimisation, however, I'm curious why it doesn't apply when key condition has monotonic function chain? 
This is quite useful, for example a table that have primary key is `dateTime` and the query with is like `SELECT * FROM TABLE WHEN toYear(key) > 2019 AND toYear(key) < 2021`. In this case, the key condition should be continuous, but the mark range lookup will fallback to exclusive search because of function `toYear`.
From my point of view, I cannot figure out why monotonic functions can affect the continuous of the key range.
Thanks!